fix: restructure queryMembers, queryUsers and ACItriggersettings
