+++ package/src/components/AutoCompleteInput/__tests__/AutoCompleteInput.test.js
@@ -8,7 +8,7 @@ import { useMockedApis } from '../../../mock-builders/api/useMockedApis';
 import { generateChannelResponse } from '../../../mock-builders/generator/channel';
 import { generateUser } from '../../../mock-builders/generator/user';
 import { getTestClientWithUser } from '../../../mock-builders/mock';
-import { ACITriggerSettings } from '../../../utils/utils';
+import { ACITriggerSettings } from '../../../utils/ACITriggerSettings';
 import { Chat } from '../../Chat/Chat';
 import { AutoCompleteInput } from '../AutoCompleteInput';
